+++ db/migrate/20260407181229_add_link_preview_to_messages.rb
@@ -1,0 +1,5 @@
+class AddLinkPreviewToMessages < ActiveRecord::Migration[8.1]
+  def change
+    add_column :messages, :link_preview, :jsonb
+  end
+end